# Refactor `selector-max-specificity` to leverage upstream helper