--- a/lib/rules/selector-max-specificity/index.cjs
+++ b/lib/rules/selector-max-specificity/index.cjs
@@ -77,7 +77,7 @@ const rule = (primary, secondaryOptions) => {
 						return;
 					}
 
-					if (!node.nodes) {
+					if (!node.nodes?.length) {
 						return zeroSpecificity();
 					}