--- a/lib/rules/selector-max-specificity/index.cjs
+++ b/lib/rules/selector-max-specificity/index.cjs
@@ -86,8 +86,7 @@ const rule = (primary, secondaryOptions) => {
 							// Calculate the diff between specificity with and without child nodes.
 							const entireSpecificity = selectorSpecificity.selectorSpecificity(node);
 
-							node.nodes = [];
-							const emptySpecificity = selectorSpecificity.selectorSpecificity(node);
+							const emptySpecificity = selectorSpecificity.selectorSpecificity(node.clone({ nodes: [] }));
 
 							return {
 								a: entireSpecificity.a - emptySpecificity.a,