--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -1,11 +1,6 @@
 import { compare, selectorSpecificity } from '@csstools/selector-specificity';
 import resolvedNestedSelector from 'postcss-resolve-nested-selector';
 
-import {
-	aNPlusBNotationPseudoClasses,
-	aNPlusBOfSNotationPseudoClasses,
-	linguisticPseudoClasses,
-} from '../../reference/selectors.mjs';
 import { assertNumber, isRegExp, isString } from '../../utils/validateTypes.mjs';
 import isStandardSyntaxRule from '../../utils/isStandardSyntaxRule.mjs';
 import isStandardSyntaxSelector from '../../utils/isStandardSyntaxSelector.mjs';
@@ -33,24 +28,6 @@ const meta = {
  * @returns {Specificity}
  */
 const zeroSpecificity = () => ({ a: 0, b: 0, c: 0 });
-
-/**
- * Calculate the sum of given specificities.
- *
- * @param {Specificity[]} specificities
- * @returns {Specificity}
- */
-const specificitySum = (specificities) => {
-	const sum = zeroSpecificity();
-
-	for (const { a, b, c } of specificities) {
-		sum.a += a;
-		sum.b += b;
-		sum.c += c;
-	}
-
-	return sum;
-};
 
 /** @type {import('stylelint').Rule<string>} */
 const rule = (primary, secondaryOptions) => {
@@ -82,146 +59,38 @@ const rule = (primary, secondaryOptions) => {
 		const isSelectorIgnored = (selector) =>
 			optionsMatches(secondaryOptions, 'ignoreSelectors', selector);
 
-		/**
-		 * Calculate the specificity of a simple selector (type, attribute, class, ID, or pseudos's own value).
-		 *
-		 * @param {import('postcss-selector-parser').Node} node
-		 * @returns {Specificity}
-		 */
-		const simpleSpecificity = (node) => {
-			if (isSelectorIgnored(node.toString())) {
-				return zeroSpecificity();
-			}
-
-			return selectorSpecificity(node);
-		};
-
-		/**
-		 * Calculate the specificity of the most specific direct child.
-		 *
-		 * @param {import('postcss-selector-parser').Container<string | undefined>} node
-		 * @returns {Specificity}
-		 */
-		const maxChildSpecificity = (node) =>
-			node.reduce((maxSpec, child) => {
-				const childSpecificity = nodeSpecificity(child); // eslint-disable-line no-use-before-define
-
-				return compare(childSpecificity, maxSpec) > 0 ? childSpecificity : maxSpec;
-			}, zeroSpecificity());
-
-		/**
-		 * If a `of <selector>` (`An+B of S`) is found in the specified pseudo node,
-		 * returns a copy of the pseudo node, ignoring a `of` selector (`An+B of S`).
-		 * Otherwise, returns the specified node as-is.
-		 *
-		 * @see https://drafts.csswg.org/selectors/#the-nth-child-pseudo
-		 * @param {import('postcss-selector-parser').Pseudo} pseudo
-		 * @returns {import('postcss-selector-parser').Pseudo}
-		 */
-		const ignoreOfSelectorIfAny = (pseudo) => {
-			/** @type {(node: import('postcss-selector-parser').Node | undefined) => boolean} */
-			const isOfSelector = (node) => node?.type === 'tag' && node.value === 'of';
-
-			/** @type {(node: import('postcss-selector-parser').Node | undefined) => boolean} */
-			const isSpace = (node) => node?.type === 'combinator' && node.value === ' ';
-
-			const nodes = pseudo.nodes[0]?.nodes ?? [];
-			const ofSelectorIndex = nodes.findIndex((child, i, children) => {
-				// Find ' of <selector>' nodes
-				return isSpace(child) && isOfSelector(children[i + 1]) && isSpace(children[i + 2]);
-			});
-
-			const ofSelector = nodes[ofSelectorIndex + 3];
-
-			if (!ofSelector || !ofSelector.value) return pseudo;
-
-			if (!isSelectorIgnored(ofSelector.value)) return pseudo;
-
-			const copy = pseudo.clone();
-			const rootSelector = copy.nodes[0];
-
-			if (rootSelector) {
-				// Remove ' of <selector>' nodes
-				rootSelector.nodes = rootSelector.nodes.slice(0, ofSelectorIndex);
-			}
-
-			return copy;
-		};
-
-		/**
-		 * Calculate the specificity of a pseudo selector including own value and children.
-		 *
-		 * @param {import('postcss-selector-parser').Pseudo} node
-		 * @returns {Specificity}
-		 */
-		const pseudoSpecificity = (node) => {
-			// `node.toString()` includes children which should be processed separately,
-			// so use `node.value` instead
-			const ownValue = node.value.toLowerCase();
-
-			if (ownValue === ':where') {
-				return zeroSpecificity();
-			}
-
-			let ownSpecificity;
-
-			if (isSelectorIgnored(ownValue)) {
-				ownSpecificity = zeroSpecificity();
-			} else if (aNPlusBOfSNotationPseudoClasses.has(ownValue.replace(/^:/, ''))) {
-				return selectorSpecificity(ignoreOfSelectorIfAny(node));
-			} else {
-				ownSpecificity = selectorSpecificity(node.clone({ nodes: [] }));
-			}
-
-			return specificitySum([ownSpecificity, maxChildSpecificity(node)]);
-		};
-
-		/**
-		 * @param {import('postcss-selector-parser').Node} node
-		 * @returns {boolean}
-		 */
-		const shouldSkipPseudoClassArgument = (node) => {
-			// postcss-selector-parser includes the arguments to nth-child() functions
-			// as "tags", so we need to ignore them ourselves.
-			// The fake-tag's "parent" is actually a selector node, whose parent
-			// should be the :nth-child pseudo node.
-			const parentNode = node.parent && node.parent.parent;
-
-			if (parentNode && parentNode.type === 'pseudo' && parentNode.value) {
-				const pseudoClass = parentNode.value.toLowerCase().replace(/^:/, '');
-
-				return (
-					aNPlusBNotationPseudoClasses.has(pseudoClass) || linguisticPseudoClasses.has(pseudoClass)
-				);
-			}
-
-			return false;
-		};
-
-		/**
-		 * Calculate the specificity of a node parsed by `postcss-selector-parser`.
-		 *
-		 * @param {import('postcss-selector-parser').Node} node
-		 * @returns {Specificity}
-		 */
-		const nodeSpecificity = (node) => {
-			if (shouldSkipPseudoClassArgument(node)) {
-				return zeroSpecificity();
-			}
-
+		/** @type {(node: import('postcss-selector-parser').Node) => Specificity | void} */
+		const isSelectorNodeIgnored = (node) => {
 			switch (node.type) {
 				case 'attribute':
 				case 'class':
 				case 'id':
 				case 'tag':
-					return simpleSpecificity(node);
-				case 'pseudo':
-					return pseudoSpecificity(node);
-				case 'selector':
-					// Calculate the sum of all the direct children
-					return specificitySum(node.map((n) => nodeSpecificity(n)));
-				default:
+					if (!isSelectorIgnored(node.toString())) {
+						return;
+					}
+
 					return zeroSpecificity();
+				case 'pseudo': {
+					if (!isSelectorIgnored(node.value.toLowerCase())) {
+						return;
+					}
+
+					if (!node.nodes) {
+						return zeroSpecificity();
+					}
+
+					const entireSpecificity = selectorSpecificity(node);
+
+					node.nodes = [];
+					const emptySpecificity = selectorSpecificity(node);
+
+					return {
+						a: entireSpecificity.a - emptySpecificity.a,
+						b: entireSpecificity.b - emptySpecificity.b,
+						c: entireSpecificity.c - emptySpecificity.c,
+					};
+				}
 			}
 		};
 
@@ -251,7 +120,14 @@ const rule = (primary, secondaryOptions) => {
 					if (!selectorTree) continue;
 
 					// Check if the selector specificity exceeds the allowed maximum
-					if (compare(maxChildSpecificity(selectorTree), maxSpecificity) > 0) {
+					if (
+						compare(
+							selectorSpecificity(selectorTree, {
+								customSpecificity: isSelectorNodeIgnored,
+							}),
+							maxSpecificity,
+						) > 0
+					) {
 						report({
 							ruleName,
 							result,