--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -78,6 +78,8 @@ const rule = (primary, secondaryOptions) => {
 						return zeroSpecificity();
 					}
 
+					// We only ignore the current pseudo-class, not the specificity of the child nodes.
+					// Calculate the diff between specificity with and without child nodes.
 					const entireSpecificity = selectorSpecificity(node);
 
 					node.nodes = [];
@@ -109,7 +111,9 @@ const rule = (primary, secondaryOptions) => {
 					if (
 						compare(
 							selectorSpecificity(resolvedSelector, {
-								customSpecificity: isSelectorNodeIgnored,
+								customSpecificity: secondaryOptions?.ignoreSelectors
+									? isSelectorNodeIgnored
+									: undefined,
 							}),
 							maxSpecificity,
 						) > 0