--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -74,7 +74,7 @@ const rule = (primary, secondaryOptions) => {
 						return;
 					}
 
-					if (!node.nodes) {
+					if (!node.nodes?.length) {
 						return zeroSpecificity();
 					}