--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -57,42 +57,46 @@ const rule = (primary, secondaryOptions) => {
 		const isSelectorIgnored = (selector) =>
 			optionsMatches(secondaryOptions, 'ignoreSelectors', selector);
 
-		/** @type {(node: import('postcss-selector-parser').Node) => Specificity | void} */
-		const isSelectorNodeIgnored = (node) => {
-			switch (node.type) {
-				case 'attribute':
-				case 'class':
-				case 'id':
-				case 'tag':
-					if (!isSelectorIgnored(node.toString())) {
-						return;
+		/** @type {import('@csstools/selector-specificity').CustomSpecificityCallback | undefined} */
+		const customSpecificity = secondaryOptions?.ignoreSelectors
+			? (node) => {
+					switch (node.type) {
+						case 'attribute':
+						case 'class':
+						case 'id':
+						case 'tag':
+							if (!isSelectorIgnored(node.toString())) {
+								return;
+							}
+
+							return zeroSpecificity();
+						case 'pseudo': {
+							if (!isSelectorIgnored(node.value.toLowerCase())) {
+								return;
+							}
+
+							if (!node.nodes?.length) {
+								return zeroSpecificity();
+							}
+
+							// We only ignore the current pseudo-class, not the specificity of the child nodes.
+							// Calculate the diff between specificity with and without child nodes.
+							const entireSpecificity = selectorSpecificity(node);
+
+							node.nodes = [];
+							const emptySpecificity = selectorSpecificity(node);
+
+							return {
+								a: entireSpecificity.a - emptySpecificity.a,
+								b: entireSpecificity.b - emptySpecificity.b,
+								c: entireSpecificity.c - emptySpecificity.c,
+							};
+						}
+						default:
+						// Other node types are not ignorable.
 					}
-
-					return zeroSpecificity();
-				case 'pseudo': {
-					if (!isSelectorIgnored(node.value.toLowerCase())) {
-						return;
-					}
-
-					if (!node.nodes?.length) {
-						return zeroSpecificity();
-					}
-
-					// We only ignore the current pseudo-class, not the specificity of the child nodes.
-					// Calculate the diff between specificity with and without child nodes.
-					const entireSpecificity = selectorSpecificity(node);
-
-					node.nodes = [];
-					const emptySpecificity = selectorSpecificity(node);
-
-					return {
-						a: entireSpecificity.a - emptySpecificity.a,
-						b: entireSpecificity.b - emptySpecificity.b,
-						c: entireSpecificity.c - emptySpecificity.c,
-					};
-				}
-			}
-		};
+			  }
+			: undefined;
 
 		const [a, b, c] = primary.split(',').map((s) => Number.parseFloat(s));
 
@@ -107,17 +111,10 @@ const rule = (primary, secondaryOptions) => {
 
 			flattenNestedSelectorsForRule(ruleNode, result).forEach(({ selector, resolvedSelectors }) => {
 				resolvedSelectors.forEach((resolvedSelector) => {
+					const specificity = selectorSpecificity(resolvedSelector, { customSpecificity });
+
 					// Check if the selector specificity exceeds the allowed maximum
-					if (
-						compare(
-							selectorSpecificity(resolvedSelector, {
-								customSpecificity: secondaryOptions?.ignoreSelectors
-									? isSelectorNodeIgnored
-									: undefined,
-							}),
-							maxSpecificity,
-						) > 0
-					) {
+					if (compare(specificity, maxSpecificity) > 0) {
 						const index = selector.first?.sourceIndex ?? 0;
 						const selectorStr = selector.toString().trim();