--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -83,8 +83,7 @@ const rule = (primary, secondaryOptions) => {
 							// Calculate the diff between specificity with and without child nodes.
 							const entireSpecificity = selectorSpecificity(node);
 
-							node.nodes = [];
-							const emptySpecificity = selectorSpecificity(node);
+							const emptySpecificity = selectorSpecificity(node.clone({ nodes: [] }));
 
 							return {
 								a: entireSpecificity.a - emptySpecificity.a,