--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -92,7 +92,8 @@ const rule = (primary, secondaryOptions) => {
 							};
 						}
 						default:
-						// Other node types are not ignorable.
+							// Other node types are not ignorable.
+							return;
 					}
 			  }
 			: undefined;