--- a/lib/rules/selector-max-specificity/index.mjs
+++ b/lib/rules/selector-max-specificity/index.mjs
@@ -92,8 +92,7 @@ const rule = (primary, secondaryOptions) => {
 							};
 						}
 						default:
-							// Other node types are not ignorable.
-							return;
+						// Other node types are not ignorable.
 					}
 			  }
 			: undefined;