--- a/package.json
+++ b/package.json
@@ -154,7 +154,7 @@
     "@csstools/css-parser-algorithms": "^2.6.3",
     "@csstools/css-tokenizer": "^2.3.1",
     "@csstools/media-query-list-parser": "^2.1.11",
-    "@csstools/selector-specificity": "^3.0.3",
+    "@csstools/selector-specificity": "^3.1.0",
     "@dual-bundle/import-meta-resolve": "^4.1.0",
     "balanced-match": "^2.0.0",
     "colord": "^2.9.3",