<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

None

> Is there anything in the PR that needs further explanation?

While working on `selector-max-specificity` for the end position fix I noticed that all of the complexity in the rule is for the `ignoreSelectors` secondary option.

Because this option is present it is needed to manually walk the selector and re-implement large sections of `@csstools/selector-specificity` so that ignoring selectors works as expected.

I've added an option upstream to customize the specificity calculations.
See : https://github.com/csstools/postcss-plugins/blob/main/packages/selector-specificity/docs/selector-specificity.selectorspecificity.md

Because this is integrated into `@csstools/selector-specificity` we can offload a lot more complexity to that package.

A good example was all of the logic around `:nth-child(n of .foo)`.
All this logic already existed upstream but couldn't be used before.

-------

A side effect of this change is that this rule is also a bit faster.

Without secondary options.

Main:
```
node scripts/benchmark-rule.mjs selector-max-specificity '10,10,10'
Warnings: 0
Mean: 370.16875366666665 ms
Deviation: 16.015808005438103 ms
```

This change:
```
node scripts/benchmark-rule.mjs selector-max-specificity '10,10,10'
Warnings: 0
Mean: 301.923139 ms
Deviation: 13.039688509741097 ms
```

With secondary options.

Main:
```
node scripts/benchmark-rule.mjs selector-max-specificity '["10,10,10", { "ignoreSelectors": "foo" }]'
Warnings: 0
Mean: 373.55708683333336 ms
Deviation: 17.037700655785976 ms
```

This change:
```
node scripts/benchmark-rule.mjs selector-max-specificity '["10,10,10", { "ignoreSelectors": "foo" }]'
Warnings: 0
Mean: 317.46817016666665 ms
Deviation: 19.0016953176133 ms
```