@Mouvedia 2024-05-12
> A _side effect_ of this change is that this rule is also a bit faster.

Be bold and say that's what you intended all along :)